requirements.txt
Flask==3.0.3
gunicorn==22.0.0
twilio==9.2.3
python-dotenv==1.0.1
groq==0.12.0

google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-httplib2==0.2.0

requests==2.32.3
pytesseract==0.3.10
Pillow==10.4.0

httpx==0.27.2
